+++ go.mod
@@ -7,9 +7,12 @@ require (
 	github.com/elazarl/goproxy v0.0.0-20220417044921-416226498f94 // indirect
 	github.com/go-openapi/swag v0.21.1 // indirect
 	github.com/grafana/grafana-plugin-sdk-go v0.134.0
-	github.com/jinzhu/copier v0.3.2
+	github.com/jinzhu/copier v0.3.5
 	github.com/magefile/mage v1.13.0 // indirect
 	github.com/patrickmn/go-cache v2.1.0+incompatible
-	golang.org/x/oauth2 v0.8.0
-	google.golang.org/api v0.124.0
+	golang.org/x/net v0.10.0 // indirect
+	golang.org/x/oauth2 v0.6.0
+	google.golang.org/api v0.114.0
+	google.golang.org/genproto v0.0.0-20230410155749-daa745c078e1 // indirect
+	google.golang.org/grpc v1.55.0 // indirect
 )
