+++ pkg/analytics.go
@@ -5,7 +5,8 @@ import (
 
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	"github.com/grafana/grafana-plugin-sdk-go/data"
-	"setting"
+	"github.com/blackcowmoo/grafana-google-analytics-dataSource/pkg/setting"
+	"github.com/blackcowmoo/grafana-google-analytics-dataSource/pkg/model"
 )
 
 type GoogleAnalytics interface {
@@ -15,6 +16,6 @@ type GoogleAnalytics interface {
 	GetProfiles(ctx context.Context, config *setting.DatasourceSecretSettings, accountId string, webPropertyId string) (map[string]string, error)
 	GetProfileTimezone(ctx context.Context, config *setting.DatasourceSecretSettings, accountId string, webPropertyId string, profileId string) (string, error)
 	GetAllProfilesList(ctx context.Context, config *setting.DatasourceSecretSettings) (map[string]string, error)
-	GetDimensions() ([]interface{}, error)
-	GetMetrics() ([]interface{}, error)
+	GetDimensions() ([]model.MetadataItem, error)
+	GetMetrics() ([]model.MetadataItem, error)
 }
