+++ pkg/util/util.go
@@ -1,4 +1,4 @@
-package main
+package util
 
 import (
 	"github.com/grafana/grafana-plugin-sdk-go/backend/log"
